# internal/swagger/swagger.yaml
definitions:
  errors.Error:
    properties:
      code:
        type: integer
      detail:
        type: string
      id:
        type: string
      status:
        type: string
    type: object
...
  schema.LoginForm:
    properties:
      captcha_code:
        description: Captcha verify code
        type: string
      captcha_id:
        description: Captcha verify id
        type: string
      password:
        description: Login password (md5 hash)
        type: string
      username:
        description: Login name
        type: string
    required:
    - captcha_code
    - captcha_id
    - password
    - username
    type: object
  schema.LoginToken:
    properties:
      access_token:
        description: Access token (JWT)
        type: string
      expires_at:
        description: 'Expired time (Unit: second)'
        type: integer
      token_type:
        description: 'Token type (Usage: Authorization=${token_type} ${access_token})'
        type: string
    type: object
  schema.Menu:
    properties:
      children:
        description: Child menus
        items:
          $ref: '#/definitions/schema.Menu'
        type: array
      code:
        description: Code of menu (unique for each level)
        type: string
      created_at:
        description: Create time
        type: string
      description:
        description: Details about menu
        type: string
      id:
        description: Unique ID
        type: string
      name:
        description: Display name of menu
        type: string
      parent_id:
        description: Parent ID (From Menu.ID)
        type: string
      parent_path:
        description: Parent path (split by .)
        type: string
      path:
        description: Access path of menu
        type: string
      properties:
        description: Properties of menu (JSON)
        type: string
      resources:
        description: Resources of menu
        items:
          $ref: '#/definitions/schema.MenuResource'
        type: array
      sequence:
        description: Sequence for sorting (Order by desc)
        type: integer
      status:
        description: Status of menu (enabled, disabled)
        type: string
      type:
        description: Type of menu (page, button)
        type: string
      updated_at:
        description: Update time
        type: string
    type: object
  schema.MenuForm:
    properties:
      code:
        description: Code of menu (unique for each level)
        maxLength: 32
        type: string
      description:
        description: Details about menu
        type: string
      name:
        description: Display name of menu
        maxLength: 128
        type: string
      parent_id:
        description: Parent ID (From Menu.ID)
        type: string
      path:
        description: Access path of menu
        type: string
      properties:
        description: Properties of menu (JSON)
        type: string
      resources:
        description: Resources of menu
        items:
          $ref: '#/definitions/schema.MenuResource'
        type: array
      sequence:
        description: Sequence for sorting (Order by desc)
        type: integer
      status:
        description: Status of menu (enabled, disabled)
        enum:
        - disabled
        - enabled
        type: string
      type:
        description: Type of menu (page, button)
        enum:
        - page
        - button
        type: string
    required:
    - code
    - name
    - status
    - type
    type: object
  schema.MenuResource:
    properties:
      created_at:
        description: Create time
        type: string
      id:
        description: Unique ID
        type: string
      menu_id:
        description: From Menu.ID
        type: string
      method:
        description: HTTP method
        type: string
      path:
        description: API request path (e.g. /api/v1/users/:id)
        type: string
      updated_at:
        description: Update time
        type: string
    type: object
...
  schema.UpdateCurrentUser:
    properties:
      email:
        description: Email of user
        maxLength: 128
        type: string
      name:
        description: Name of user
        maxLength: 64
        type: string
      phone:
        description: Phone number of user
        maxLength: 32
        type: string
      remark:
        description: Remark of user
        maxLength: 1024
        type: string
    required:
    - name
    type: object
  schema.UpdateLoginPassword:
    properties:
      new_password:
        description: New password (md5 hash)
        type: string
      old_password:
        description: Old password (md5 hash)
        type: string
    required:
    - new_password
    - old_password
    type: object
  schema.User:
    properties:
      created_at:
        description: Create time
        type: string
      email:
        description: Email of user
        type: string
      id:
        description: Unique ID
        type: string
      name:
        description: Name of user
        type: string
      phone:
        description: Phone number of user
        type: string
      remark:
        description: Remark of user
        type: string
      roles:
        description: Roles of user
        items:
          $ref: '#/definitions/schema.UserRole'
        type: array
      status:
        description: Status of user (activated, freezed)
        type: string
      updated_at:
        description: Update time
        type: string
      username:
        description: Username for login
        type: string
    type: object
...
  schema.UserRole:
    properties:
      created_at:
        description: Create time
        type: string
      id:
        description: Unique ID
        type: string
      role_id:
        description: From Role.ID
        type: string
      role_name:
        description: From Role.Name
        type: string
      updated_at:
        description: Update time
        type: string
      user_id:
        description: From User.ID
        type: string
    type: object
  util.ResponseResult:
    properties:
      data: {}
      error:
        $ref: '#/definitions/errors.Error'
      success:
        type: boolean
      total:
        type: integer
    type: object
info:
  contact: {}
  description: A test API service based on golang.
  title: gastarter
  version: v1.0.0
paths:
...
  /api/v1/captcha/image:
    get:
      parameters:
      - description: Captcha ID
        in: query
        name: id
        required: true
        type: string
      - description: Reload captcha image (reload=1)
        in: query
        name: reload
        type: number
      produces:
      - image/png
      responses:
        "200":
          description: Captcha image
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ResponseResult'
      summary: Response captcha image
      tags:
      - LoginAPI
  /api/v1/current/logout:
    post:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Logout system
      tags:
      - LoginAPI
  /api/v1/current/menus:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.Menu'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Query current user menus based on the current user role
      tags:
      - LoginAPI
  /api/v1/current/password:
    put:
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.UpdateLoginPassword'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Change current user password
      tags:
      - LoginAPI
  /api/v1/current/refresh-token:
    post:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.LoginToken'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Refresh current access token
      tags:
      - LoginAPI
  /api/v1/current/user:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.User'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Get current user info
      tags:
      - LoginAPI
    put:
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.UpdateCurrentUser'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Update current user info
      tags:
      - LoginAPI
...
  /api/v1/login:
    post:
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.LoginForm'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.LoginToken'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      summary: Login system with username and password
      tags:
      - LoginAPI
  /api/v1/menus:
    get:
      parameters:
      - description: Code path of menu (like xxx.xxx.xxx)
        in: query
        name: code
        type: string
      - description: Name of menu
        in: query
        name: name
        type: string
      - description: Whether to include menu resources
        in: query
        name: includeResources
        type: boolean
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.Menu'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Query menu tree data
      tags:
      - MenuAPI
    post:
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.MenuForm'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.Menu'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Create menu record
      tags:
      - MenuAPI
  /api/v1/menus/{id}:
    delete:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Delete menu record by ID
      tags:
      - MenuAPI
    get:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.Menu'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Get menu record by ID
      tags:
      - MenuAPI
    put:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.MenuForm'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Update menu record by ID
      tags:
      - MenuAPI
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"